pyproject.toml
[project]
name = "portable-agentic-toolkit"
version = "0.1.0"
description = "Portable, privacy-first AI agent workspace powered by local LLMs"
requires-python = ">=3.11"
dependencies = [
    "chainlit>=1.3.0",
    "crewai[tools]>=0.80.0",
    "fastapi>=0.111.0",
    "mcp>=1.4.0",
    "ollama>=0.4.0",
    "pyyaml>=6.0.2",
    "pydantic>=2.10.0",
    "httpx>=0.27.0",
    "litellm>=1.50.0",
    "uvicorn[standard]>=0.29.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["agents", "mcp_tools", "utils", "db", "api"]

[tool.uv]
dev-dependencies = [
    "ruff>=0.4.0",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "W", "I"]   # pycodestyle + pyflakes + isort
ignore = [
    "E501",   # line length — handled by formatter
    "I001",   # import sort — enforced locally via `ruff check --fix`, not in CI
]

[tool.ruff.lint.isort]
known-first-party = ["agents", "mcp_tools", "utils"]
